Reduce remaining program crates to boilerplate crates
