+++ programs/storage_program/Cargo.toml
@@ -9,16 +9,10 @@ homepage = "https://solana.com/"
 edition = "2018"
 
 [dependencies]
-bincode = "1.1.2"
 log = "0.4.2"
-serde = "1.0.89"
-serde_derive = "1.0.89"
 solana-logger = { path = "../../logger", version = "0.13.0" }
 solana-sdk = { path = "../../sdk", version = "0.13.0" }
 solana-storage-api = { path = "../storage_api", version = "0.13.0" }
-
-[dev-dependencies]
-solana-runtime = { path = "../../runtime", version = "0.13.0" }
 
 [lib]
 name = "solana_storage_program"
