+++ programs/token_api/Cargo.toml
@@ -10,8 +10,10 @@ edition = "2018"
 
 [dependencies]
 bincode = "1.1.2"
+log = "0.4.2"
 serde = "1.0.89"
 serde_derive = "1.0.89"
+solana-logger = { path = "../../logger", version = "0.13.0" }
 solana-sdk = { path = "../../sdk", version = "0.13.0" }
 
 [lib]
