+++ programs/token_api/src/lib.rs
@@ -1,3 +1,6 @@
+pub mod token_processor;
+mod token_state;
+
 use solana_sdk::pubkey::Pubkey;
 
 const TOKEN_PROGRAM_ID: [u8; 32] = [
