+++ programs/token_api/src/token_processor.rs
@@ -1,0 +1,20 @@
+use crate::token_state::TokenState;
+use bincode::serialize;
+use log::*;
+use solana_sdk::account::KeyedAccount;
+use solana_sdk::pubkey::Pubkey;
+use solana_sdk::transaction::InstructionError;
+
+pub fn process_instruction(
+    program_id: &Pubkey,
+    info: &mut [KeyedAccount],
+    input: &[u8],
+    _tick_height: u64,
+) -> Result<(), InstructionError> {
+    solana_logger::setup();
+
+    TokenState::process(program_id, info, input).map_err(|e| {
+        error!("error: {:?}", e);
+        InstructionError::CustomError(serialize(&e).unwrap())
+    })
+}
