+++ programs/token_program/Cargo.toml
@@ -9,12 +9,10 @@ homepage = "https://solana.com/"
 edition = "2018"
 
 [dependencies]
-bincode = "1.1.2"
 log = "0.4.2"
-serde = "1.0.89"
-serde_derive = "1.0.89"
 solana-logger = { path = "../../logger", version = "0.13.0" }
 solana-sdk = { path = "../../sdk", version = "0.13.0" }
+solana-token-api = { path = "../token_api", version = "0.13.0" }
 
 [lib]
 name = "solana_token_program"
