+++ programs/vote_api/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod vote_instruction;
+pub mod vote_processor;
 pub mod vote_state;
 pub mod vote_transaction;
