+++ programs/vote_program/Cargo.toml
@@ -9,15 +9,10 @@ homepage = "https://solana.com/"
 edition = "2018"
 
 [dependencies]
-bincode = "1.1.2"
 log = "0.4.2"
 solana-logger = { path = "../../logger", version = "0.13.0" }
-solana-metrics = { path = "../../metrics", version = "0.13.0" }
 solana-sdk = { path = "../../sdk", version = "0.13.0" }
 solana-vote-api = { path = "../vote_api", version = "0.13.0" }
-
-[dev-dependencies]
-solana-runtime = { path = "../../runtime", version = "0.13.0" }
 
 [lib]
 name = "solana_vote_program"
